+++ lib/brcobranca.rb
@@ -99,6 +99,7 @@ module Brcobranca
     autoload :Base,          'brcobranca/boleto/base'
     autoload :BancoNordeste, 'brcobranca/boleto/banco_nordeste'
     autoload :BancoBrasil,   'brcobranca/boleto/banco_brasil'
+    autoload :BancoBrasilia, 'brcobranca/boleto/banco_brasilia'
     autoload :Itau,          'brcobranca/boleto/itau'
     autoload :Hsbc,          'brcobranca/boleto/hsbc'
     autoload :Bradesco,      'brcobranca/boleto/bradesco'