+++ lib/brcobranca/retorno/cnab400/itau.rb
@@ -50,8 +50,8 @@ module Brcobranca
           # :carteira, 107..107 #código da carteira
           parse.field :carteira, 107..107
 
-          # :cod_de_ocorrencia, 108..109 # código de ocorrencia
-          # :data_de_ocorrencia, 110..115 # data de ocorrencia no banco (ddmmaa)
+          # :codigo_ocorrencia, 108..109 # código de ocorrencia
+          # :data_ocorrencia, 110..115 # data de ocorrencia no banco (ddmmaa)
           # :n_do_documento, 116..125 # n umero do documento de cobranca (dupl, np etc)
           # :nosso_numero, 126..133 # confirmacao do numero do titulo no banco
           # :brancos, 134..145 #complemento de registro